Configure webpack dev and prod
