+++ frontend/src/template.html
@@ -3,18 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./css/index.css">
-    <link rel="stylesheet" href="./css/throbber.css">
-    <link rel="stylesheet" href="./css/scroll.css">
-    <link rel="stylesheet" href="./css/viewports.css">
-    <link rel="stylesheet" href="./css/article.css">
-    <link rel="stylesheet" href="./css/header.css">
-    <link rel="stylesheet" href="./css/tabGroup.css">
     <link rel="icon" type="image/x-icon" href="./assets/favicon.ico">
     <title>Destiny Article Finder</title>
 </head>
 <body>
-
+    <!-- Hello World This is a test -->
     <div>
         <div id="headerContainer">
             <div id="logoContainer">
